requirements.txt
langchain==0.2.0
langchain-community>=0.0.21
langchain-openai>=0.1.6
openai>=1.25.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
python-dotenv>=1.0.1